requirements.txt
py-zabbix==1.1.7
google-auth==2.12.0
google-cloud-bigquery==3.3.3
pandas==2.1.0
python-dotenv==1.0.0